+++ QEfficient/exporter/weight_free/__init__.py
@@ -5,7 +5,7 @@
 #
 # ----------------------------------------------------------------------------
 
-from QEfficient.exporter.weight_free.runtime import load_weight_free_ort_inputs
-from QEfficient.exporter.weight_free.spec import resolve_weight_spec_path
+from QEfficient.exporter.weight_free.ort_weight_injection import load_weight_free_ort_inputs
+from QEfficient.exporter.weight_free.weight_spec import resolve_weight_spec_path
 
 __all__ = ["load_weight_free_ort_inputs", "resolve_weight_spec_path"]
